# internal/models/models.go
package models

import "time"

type UserRole string

const (
	RoleOwner  UserRole = "owner"
	RoleMember UserRole = "member"
)

type User struct {
	ID                 string     `json:"id"`
	Email              string     `json:"email"`
	Password           string     `json:"-"`
	FirstName          string     `json:"first_name"`
	LastName           string     `json:"last_name"`
	Role               UserRole   `json:"role"`
	TokenInvalidatedAt *time.Time `json:"-"`
	CreatedAt          time.Time  `json:"created_at"`
	UpdatedAt          time.Time  `json:"updated_at"`
}

type Category struct {
	ID          string    `json:"id"`
	Name        string    `json:"name"`
	Description string    `json:"description"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type Book struct {
	ID           string    `json:"id"`
	Title        string    `json:"title"`
	Description  string    `json:"description"`
	PDFFile      string    `json:"pdf_file"`
	CategoryID   string    `json:"category_id"`
	OwnerID      string    `json:"owner_id"`
	LikeCount    int       `json:"like_count"`
	DislikeCount int       `json:"dislike_count"`
	SaveCount    int       `json:"save_count"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}

type BookWithCategory struct {
	Book
	CategoryName string `json:"category_name"`
}

type SavedBook struct {
	ID        string    `json:"id"`
	UserID    string    `json:"user_id"`
	BookID    string    `json:"book_id"`
	CreatedAt time.Time `json:"created_at"`
}

type Comment struct {
	ID        string    `json:"id"`
	BookID    string    `json:"book_id"`
	UserID    string    `json:"user_id"`
	Content   string    `json:"content"`
	ParentID  *string   `json:"parent_id,omitempty"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type CommentWithUser struct {
	Comment
	UserFirstName string `json:"user_first_name"`
	UserLastName  string `json:"user_last_name"`
}

type Like struct {
	ID        string    `json:"id"`
	BookID    string    `json:"book_id"`
	UserID    string    `json:"user_id"`
	IsLike    bool      `json:"is_like"`
	CreatedAt time.Time `json:"created_at"`
}

type RefreshToken struct {
	ID        string    `json:"id"`
	UserID    string    `json:"user_id"`
	Token     string    `json:"token"`
	ExpiresAt time.Time `json:"expires_at"`
	CreatedAt time.Time `json:"created_at"`
}

type Download struct {
	ID           string    `json:"id"`
	BookID       string    `json:"book_id"`
	UserID       string    `json:"user_id"`
	DownloadedAt time.Time `json:"downloaded_at"`
	IPAddress    string    `json:"ip_address,omitempty"`
	UserAgent    string    `json:"user_agent,omitempty"`
}